dont use volume slider specific key names in the base function and make the function take the slider position as a double instead of an int
svn path=/trunk/extragear/multimedia/amarok/; revision=807031

+++ src/widgets/SliderWidget.h
@@ -58,7 +58,7 @@ namespace Amarok
             virtual void mousePressEvent( QMouseEvent* );
             virtual void slideEvent( QMouseEvent* );
 
-            void paintCustomSlider( QPainter *p, int x, int y, int width, int height, int pos = -1 );
+            void paintCustomSlider( QPainter *p, int x, int y, int width, int height, double pos = -1.0 );
 
             bool m_sliding;
